Refactor phase 3h: split the alert verification routes into a package
webapp/routes/alert_verification.py (1,668 lines) becomes the
webapp/routes/alert_verification/ package: 14 modules plus a 123-line
__init__, all within the guidance. Phase 3 is now complete for every
file except app.py, which is assessed separately as 3g.

687 of those lines were a single register(app, logger) with eight
helpers and seven handlers nested inside it, so this split had to
reproduce a closure rather than just move text. Which of the two kinds
of motion applied was read off symtable, never guessed: four helpers
capture nothing and were dedented to module scope in helpers.py; the
other four capture route_logger, repo_root or app, so each topic module
keeps its own register(app, logger) that rebuilds exactly the locals its
handlers need. Emitting both locals unconditionally left an unused
repo_root in three modules.

Four mutable globals are deliberately not re-exported from the package.
_progress_dir, _progress_lock, _result_dir and _result_lock are replaced
with tmp_path by the async tests. This was measured, not assumed: with
the names re-exported and the patch aimed at the package, all six tests
still pass while writing to the real temp directory. The patch rebinds a
copy; the classes keep reading the original. Absent from the package the
same call raises AttributeError naming the module. Six patch sites
across two test files were retargeted.

An import cycle was created and caught during the split. The chain
_process_temp_audio_file -> _detect_comprehensive_eas_segments ->
_build_composite_audio_segment crosses three layers; grouping its two
ends into one composite_audio module, which their names invite,
sandwiched eas_detection into a cycle. Split into composite_audio and
temp_audio; both the generator and the package test now walk the emitted
imports and fail with the cycle named.

Also fixes a pre-existing flaky test that made every "suite green" claim
unreliable: tests/test_8khz_stress_test.py generated noise with the
global unseeded random and then asserted a confidence threshold, failing
2 runs in 12. It now seeds a local random.Random. Seeding surfaced that
6 of 50 seed/noise combinations fail at 20-25% noise with zero headers
despite ~0.92 bit confidence — bit recovery healthy, header framing gone
— which is documented in the helper docstring with a reproducer rather
than hidden by the seed.

Verification: 27/28 definitions ast.dump()-identical including the nested
ones (the difference is register itself, deliberately restructured to
fan out), URL map unchanged at 549 rules / 0 differences with all 7
endpoints intact, full suite 2301 passed.
